Transactions contoller and services are added!!

+++ src/main/java/roy/aman/bankingmanagementsystem/Entity/Account.java
@@ -36,7 +36,7 @@ public abstract class Account {
 //    )
 //    private Long accountNumber;
 
-    private Long totalAmount;
+    private Integer totalAmount;
 
     private Date dateOfOpening;
